Implemented recommendation system, introducing new components: Review and CustomSelect. Refactored controllers and routes to resolve errors. Enhanced Destination details page with a new section for reviews. Improved responsiveness by adding and modifying elements.

+++ backend/controllers/reviews.js
@@ -13,14 +13,14 @@ exports.getReviews = asyncHandler(async (req, res, next) => {
   if (req.params.destinationId) {
     const reviews = await Review.find({
       destination: req.params.destinationId,
-    });
+    }).populate('user destination booking');
 
-    return res.status(200).json(res.advancedResults);
-    // return res.status(200).json({
-    //   success: true,
-    //   count: reviews.length,
-    //   data: reviews,
-    // });
+    // return res.status(200).json(res.advancedResults);
+    return res.status(200).json({
+      success: true,
+      count: reviews.length,
+      data: reviews,
+    });
   } else {
     res.status(200).json(res.advancedResults);
   }
@@ -55,17 +55,20 @@ exports.getReview = asyncHandler(async (req, res, next) => {
 exports.addReview = asyncHandler(async (req, res, next) => {
   req.body.destination = req.params.destinationId;
   req.body.user = req.user.id;
+  req.body.booking = req.body.bookingId;
 
   const destination = await Destination.findById(req.params.destinationId);
 
   if (!destination) {
     return next(
       new ErrorResponse(
-        `Destination with the id of ${req.params.destinationId} not found`
-      ),
-      404
+        `Destination with the id of ${req.params.destinationId} not found`,
+        404
+      )
     );
   }
+
+  console.log(req.body);
 
   const review = await Review.create(req.body);
 
@@ -88,7 +91,6 @@ exports.updateReview = asyncHandler(async (req, res, next) => {
     );
   }
 
-  // Make sure review belongs to user or user is admin
   if (review.user.toString() !== req.user.id && req.user.role !== 'admin') {
     return next(
       new ErrorResponse(
@@ -98,16 +100,21 @@ exports.updateReview = asyncHandler(async (req, res, next) => {
     );
   }
 
-  review = await Review.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  const { comment } = req.body;
+  review.comment = comment;
+
+  await review.save();
+
+  review = await Review.findById(req.params.id).populate(
+    'user destination booking'
+  );
 
   res.status(200).json({
     success: true,
     data: review,
   });
 });
+
 // @desc Delete review
 // @route DELETE /reviews/:id
 // @access Private
@@ -131,11 +138,34 @@ exports.deleteReview = asyncHandler(async (req, res, next) => {
     );
   }
 
-  // await review.deleteOne();
+  const destination = await Destination.findById(review.destination._id);
+
+  if (!destination) {
+    return next(
+      new ErrorResponse(`Destination associated with the review not found`, 404)
+    );
+  }
+
+  // Delete review
   await Review.findByIdAndDelete(req.params.id);
+
+  // Recalculate average rating
+  const reviews = await Review.find({ destination: destination._id });
+  const totalReviews = reviews.length;
+  const totalRating = reviews.reduce((acc, curr) => acc + curr.rating, 0);
+  destination.averageRating =
+    totalReviews > 0 ? totalRating / totalReviews : undefined;
+
+  await destination.save();
+
+  if (req.params.destinationId) {
+    const reviews = await Review.find({
+      destination: req.params.destinationId,
+    });
+  }
 
   res.status(200).json({
     success: true,
-    data: { id: review._id },
+    data: { id: review._id, reviews: reviews },
   });
 });
